refactor for cloud->template->instance
